requirements.txt
tensorflow>=2.10
torch>=2.0
transformers>=4.35
pillow
numpy
matplotlib
scikit-learn
tqdm